requirements.txt
Django==1.4.2
django-grappelli==2.4.2
django-guardian==1.0.4
django-userena==1.1.2
easy-thumbnails==1.1
wsgiref==0.1.2
